feat: 系统设置页面初始化（框架搭建）
