+++ src/renderer/src/components/Setting/index.jsx
@@ -1,0 +1,11 @@
+import styles from './styles.module.scss'
+
+const Setting = () => {
+	return (
+		<div className={styles['setting']}>
+			<div>11111111111</div>
+		</div>
+	)
+}
+
+export default Setting
