+++ src/renderer/src/store/index.js
@@ -1,8 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit'
 import chatSlice from './slice/chatSlice'
+import systemSlice from './slice/systemSlice'
 
 export default configureStore({
 	reducer: {
-		chat: chatSlice
+		chat: chatSlice,
+		system: systemSlice
 	}
 })
